scripts/DNN_training/requirements.txt
appdirs==1.4.3
audioread==2.1.5
backports.weakref==1.0rc1
bleach==1.5.0
cycler==0.10.0
Cython==0.25.2
daemonize==2.4.7
decorator==4.0.6
editdistance==0.3.1
funcsigs==1.0.2
functools32==3.2.3.post2
graphviz==0.7.1
guppy==0.1.10
h5py==2.7.0
htk-io==0.5
html5lib==0.9999999
ipython==2.4.1
joblib==0.11
Keras==2.0.6
Lasagne==0.2.dev1
librosa==0.5.1
Mako==1.0.6
Markdown==2.2.0
MarkupSafe==1.0
matplotlib==2.0.2
memory-profiler==0.47
mock==2.0.0
nose==1.3.7
numpy==1.13.1
packaging==16.8
pbr==3.1.1
pexpect==4.0.1
posix-ipc==1.0.0
protobuf==3.3.0
ptyprocess==0.5
py==1.4.33
pycurl==7.43.0
pydot==1.2.3
pydot-ng==1.0.0
pyfst==0.2.3
pygpu==0.6.5
pyliblzma==0.5.3
pyparsing==2.2.0
pysqlite==1.0.1
pytest==3.0.7
python-apt==1.1.0b1
python-dateutil==2.6.0
python-speech-features==0.5
pytools==2016.2.6
pytz==2017.2
PyYAML==3.12
resampy==0.1.5
rpm-python==4.12.0.1
scikit-learn==0.18.1
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
subprocess32==3.2.7
tensorflow-gpu==1.2.0
tfrbm==0.0.2
Theano==0.9.0
tkinter==0.2.0
tqdm==4.14.0
urlgrabber==3.9.1
virtualenv==15.0.1
Werkzeug==0.12.2
yum-metadata-parser==1.1.4
